\name{decomppath}
\alias{decomppath}
\docType{data}
\title{Demonstration Decomposition Path Matrix}
\description{A decomposition path matrix for use with the supplied demonstration data.  It is inteded for use for comparing two 16x16 pixel categorical maps along a spectra of spatial decomposition.}
\usage{data(decomppath)}
\examples{
data(decomppath)
str(decomppath)
decomppath
}
\keyword{datasets}
