\name{demoimage1}
\alias{demoimage1}
\docType{data}
\title{Demonstration binary 16x16 categorical map 1}
\description{For use with hdeco (and other functions) to illustrate usage.  This categorical map has two vertical bars.}
\usage{data(demoimage1)}
\examples{
data(demoimage1)
str(demoimage1)
fixedcolimage(demoimage1)
}
\keyword{datasets}
