\name{demoimage2}
\alias{demoimage2}
\docType{data}
\title{Demonstration binary 16x16 categorical map 2}
\description{For use with hdeco to illustrate usage.  This categorical map has 2 horizontal bars.}
\usage{data(demoimage1)}
\examples{
data(demoimage2)
str(demoimage2)
fixedcolimage(demoimage2)
}
\keyword{datasets}
