\name{hdeco-package}
\alias{hdeco-package}
\docType{package}
\title{Hierarchical DECOmposition of Entropy for Categorical Map Comparisons}
\description{We provide a measurement scheme for the comparison of categorical maps that decomposes the differences in multidimensional nested coincidence tables according to variables that record occurrence frequencies of categories (Z), at levels of spatial aggregation (Y), on specific maps (X). Sequences of conditional entropies computed according to the specific questions asked (e.g. is there coincidence between colours and locations), characterize the correspondence between the three types of variables in common units (bits) measured by mutual information. The form of these sequences, as a variable runs from coarse to fine detail, referred to as spectra, provide meaningful characterizations of the similarities/differences between categorical maps, including their spatial configuration.}
\details{
\tabular{ll}{
Package: \tab hdeco\cr
Type: \tab Package\cr
Version: \tab 0.4\cr
Date: \tab 2007-05-02\cr
License: \tab GPL (Version 2.0 or later)\cr
}
}
\author{Tarmo K. Remmel and Sandor Kabos\cr
Maintainer: Tarmo K. Remmel <remmelt@yorku.ca>}
\references{
  Remmel, T.K. and F. Csillag. 2006. Mutual information spectra for comparing categorical maps. International Journal of Remote Sensing 27(7)1425-1452.   
  \cr\cr
  Remmel, T.K., F. Csillag, and S. Kabos. 2004. Using hierarchical entropy decomposition to compare the spatial patterns found in categorical digital maps. In Abstracts of the Canadian Association of Geographers (CAG) Annual Meeting, May 25-29, Moncton, New Brunswick, Canada.
}
\keyword{package}
