\name{HPROFIL}
\alias{HPROFIL}
\docType{data}
\title{Sample HDECO output data}
\description{A sample .HPROFIL matrix renamed to HPROFIL that can be used to test the plotting functions of sigplot etc. }
\usage{data(HPROFIL)}
\seealso{See Also \code{\link{sigplot}}, \code{\link{hdeco}} }
\examples{
data(HPROFIL)
str(HPROFIL)
HPROFIL
}
\keyword{datasets}
