\name{ls.}
\alias{ls.}
\title{List objects starting with '.'}
\description{Similar to the function ls, this modified form shows only those objects whose names begin with a period.}
\usage{ls.()}
\details{This function will only display the hidden objects in a workspace.  To see other objects, use the traditional form of the list function, ls}
\value{Provides a listing of all objects in the workspace whose names begin with a period.}
\author{Ferenc Csillag}
\note{Useful for listing hidden objects in a workspace that are generated by HDECO}
\seealso{\code{\link{hdeco}}, \code{\link{ls}}}
\examples{ls.()}
\keyword{misc}
